requirements.txt
pywebview
BeautifulSoup4